+++ CHANGELOG.md
@@ -45,6 +45,7 @@
   - Every 10 minutes, the server invalidations are retrieved to compare local invalidations with server ones.
   - Add `store.isAutoRequest` atttribute to enable or disable automatic `lunaris.get` call on `reset` event when subscribing to store with `vm.stores`.
   - Add `lunaris.collectionResultSet` to query collection data (more in documentation) !.
+  - Add `dependencies` option for `lunaris.build`.
 - Fix:
   - Improve offline app.
   - In offline or offline mode, generated urls were not encoded.