+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Lunaris
 
 ## 1.15.0
+*2020-01-06*
   - Add `lunaris.offline.isOfflineMode` attribute to set offline mode. In online the offline mode capture the invalidations and simulate offline connection.
   - Add `lunaris.offline.isSynchronizing` attribute. Internal use.
   - Add `lunaris._onInvalidate(event, handler)` to capture events in invalidation module. Only `invalidate` event is sent. It allows an handler to capture invalidations when offline mode is activated.